requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.32.2
python-dotenv